requirements.txt
flask
gunicorn 
psycopg2-binary
PyYAML
feedparser
wget
requests
